unsigned_long_int var populacao

+++ supertrunfo.c
@@ -6,7 +6,7 @@ int main (){
     char estado [] = "B";
     char codigo [] = "B01";
     char cidade [] = "Salvador";
-    int populacao = 2564204;
+    unsigned long int populacao = 2564204;
     float area = 692.9;
     float pib = 73.5; // em bilhões de reais
     int pontosTuristicos = 20;
@@ -15,7 +15,7 @@ int main (){
     float pibPerCapita = (pib * 1000000000.0f) / (float)populacao; // PIB em reais dividido pela população
     float superPoder = (float)populacao + area + pib + densidade + densidadeInvertida + pibPerCapita;
 
-    printf("Carta 1: \n Estado: %s; \n Codigo: %s; \n Cidade: %s; \n Populacao: %d; \n " 
+    printf("Carta 1: \n Estado: %s; \n Codigo: %s; \n Cidade: %s; \n Populacao: %lu; \n " 
     "Area: %.2f km2; \n PIB: %.1f bilhoes de reais; \n Pontos Turisticos: %d; \n Densidade "
     "Populacional: %.2f hab/km2; \n Pib Per Capita: %.2f reais; \n Super Poder: %.2f. \n",
     estado, codigo, cidade, populacao, area, pib, pontosTuristicos, densidade, pibPerCapita, superPoder);
@@ -24,7 +24,7 @@ int main (){
     char estado2 [] = "A";
     char codigo2 [] = "A01";
     char cidade2 [] = "Aracaju";
-    int populacao2 = 630932;
+    unsigned long int populacao2 = 630932;
     float area2 = 182.16;
     float pib2 = 16.5; // em bilhões de reais
     int pontosTuristicos2 = 32;
@@ -33,7 +33,7 @@ int main (){
     float pibPerCapita2 = (pib2 * 1000000000.0f) / (float)populacao2; // PIB em reais dividido pela população
     float superPoder2 = (float)populacao2 + area2 + pib2 + densidade2 + densidadeInvertida2 + pibPerCapita2;
 
-    printf("Carta 2: \n Estado: %s; \n Codigo: %s; \n Cidade: %s; \n Populacao: %d; \n " 
+    printf("Carta 2: \n Estado: %s; \n Codigo: %s; \n Cidade: %s; \n Populacao: %lu; \n " 
     "Area: %.2f km2; \n PIB: %.1f bilhoes de reais; \n Pontos Turisticos: %d; \n Densidade "
     "Populacional: %.2f hab/km2; \n Pib Per Capita: %.2f reais; \n Super Poder: %.2f. \n",
     estado2, codigo2, cidade2, populacao2, area2, pib2, pontosTuristicos2, densidade2, pibPerCapita2, superPoder2);
